Updated to bevy 0.17

+++ src/core/constants.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 
-pub const WINDOW_WIDTH: f32 = 1280.0;
-pub const WINDOW_HEIGHT: f32 = 720.0;
+pub const WINDOW_WIDTH: u32 = 1280;
+pub const WINDOW_HEIGHT: u32 = 720;
 
 pub const NUM_GHOST_TYPES: usize = 4;
